fix : ExcelPerformOperations.java move updatedocument() in line 149

+++ src/main/java/com/chicmic/engine/FolderHeirarchyAndFourPoint.java
@@ -23,6 +23,7 @@ public class FolderHeirarchyAndFourPoint {
             fourPointDeclaration.generateDocument(excelFile);
 
             // Delete Temp Files
+            if(MainRunner.deleteTempFiles)
             folderOperations.deleteFile(MainRunner.rootDirectory + "/sorted_" + MainRunner.FILE_NAME);
 
         } catch (IOException e) {
